make intervalTypes a const - move constants folder to be more central - rename import file paths - add INTERAL_STYLES const - change "---" to "--" - use INTERVAL_STYLES throughout
