+++ src/constants/Chart-constants.ts
@@ -1,0 +1,1 @@
+export const INTERVAL_STYLES = ["--", "intervals", "error bars"] as const;
